requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.6.0
async-timeout==4.0.3
billiard==3.6.4.0
blis==1.0.1
catalogue==2.0.10
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
confection==0.1.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cymem==2.0.8
defusedxml==0.7.1
Django==4.1.7
django-celery-beat==2.4.0
django-cors-headers==4.5.0
django-filter==22.1
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.5
filelock==3.16.1
fsspec==2024.10.0
gunicorn==20.1.0
huggingface-hub==0.26.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.4.2
keybert==0.8.5
kombu==5.4.2
langcodes==3.4.1
language_data==1.2.0
language_tool_python==2.8.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
pillow==11.0.0
preshed==3.0.9
prompt_toolkit==3.0.47
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.2
redis==4.5.1
redis-cli==0.1.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.9.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
shellingham==1.5.4
simplejson==3.18.3
six==1.16.0
smart-open==7.0.5
social-auth-app-django==4.0.0
social-auth-core==4.3.0
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.3
srsly==2.4.8
stripe==5.2.0
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
